python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.26.4
opencv-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
